+++ src/Mapper.js
@@ -13,6 +13,10 @@ import {
 } from './decorators'
 
 const DOMAIN = 'Mapper'
+const applyDefaultsHooks = [
+  'beforeCreate',
+  'beforeCreateMany'
+]
 const validatingHooks = [
   'beforeCreate',
   'beforeCreateMany',
@@ -26,8 +30,21 @@ const makeNotify = function (num) {
     const op = opts.op
     this.dbg(op, ...args)
 
+    if (applyDefaultsHooks.indexOf(op) !== -1 && opts.applyDefaults !== false) {
+      const schema = this.getSchema()
+      if (schema && schema.applyDefaults) {
+        let toProcess = args[0]
+        if (!utils.isArray(toProcess)) {
+          toProcess = [toProcess]
+        }
+        toProcess.forEach((record) => {
+          schema.applyDefaults(record)
+        })
+      }
+    }
+
     // Automatic validation
-    if (validatingHooks.indexOf(op) !== -1 && opts.validate !== false) {
+    if (validatingHooks.indexOf(op) !== -1 && !opts.noValidate) {
       // Save current value of option
       const originalExistingOnly = opts.existingOnly
 
@@ -131,6 +148,17 @@ const MAPPER_DEFAULTS = {
   _adapters: {},
 
   /**
+   * Whether {@link Mapper#beforeCreate} and {@link Mapper#beforeCreateMany}
+   * should automatically receive default values according to the Mapper's schema.
+   *
+   * @default true
+   * @name Mapper#applyDefaults
+   * @since 3.0.0
+   * @type {boolean}
+   */
+  applyDefaults: true,
+
+  /**
    * Whether to augment {@link Mapper#recordClass} with ES5 getters and setters
    * according to the properties defined in {@link Mapper#schema}. This makes
    * possible validation and change tracking on individual properties
@@ -174,6 +202,16 @@ const MAPPER_DEFAULTS = {
    * @type {boolean}
    */
   notify: true,
+
+  /**
+   * Whether to skip validation when the Record instances are created.
+   *
+   * @default false
+   * @name Mapper#noValidate
+   * @since 3.0.0
+   * @type {boolean}
+   */
+  noValidate: false,
 
   /**
    * Whether {@link Mapper#create}, {@link Mapper#createMany},
@@ -342,7 +380,7 @@ function Mapper (opts) {
      *
      * @example <caption>Mapper#schema</caption>
      * // Normally you would do: import {Mapper} from 'js-data'
-     * const JSData = require('js-data@3.0.0-beta.7')
+     * const JSData = require('js-data@3.0.0-beta.10')
      * const {Mapper} = JSData
      * console.log('Using JSData v' + JSData.version.full)
      *
@@ -897,6 +935,7 @@ export default Component.extend({
    * of whatever adapter you're using for more configuration options.
    * @param {boolean} [opts.adapter={@link Mapper#defaultAdapter}] Name of the
    * adapter to use.
+   * @param {boolean} [opts.noValidate={@link Mapper#noValidate}] See {@link Mapper#noValidate}.
    * @param {boolean} [opts.notify={@link Mapper#notify}] See {@link Mapper#notify}.
    * @param {boolean} [opts.raw={@link Mapper#raw}] See {@link Mapper#raw}.
    * @param {string[]} [opts.with=[]] Relations to create in a cascading
@@ -1104,6 +1143,7 @@ export default Component.extend({
    * method of whatever adapter you're using for more configuration options.
    * @param {boolean} [opts.adapter={@link Mapper#defaultAdapter}] Name of the
    * adapter to use.
+   * @param {boolean} [opts.noValidate={@link Mapper#noValidate}] See {@link Mapper#noValidate}.
    * @param {boolean} [opts.notify={@link Mapper#notify}] See {@link Mapper#notify}.
    * @param {boolean} [opts.raw={@link Mapper#raw}] See {@link Mapper#raw}.
    * @param {string[]} [opts.with=[]] Relations to create in a cascading
@@ -1295,8 +1335,7 @@ export default Component.extend({
    * @param {Object|Object[]} props The properties for the Record instance or an
    * array of property objects for the Record instances.
    * @param {Object} [opts] Configuration options.
-   * @param {boolean} [opts.noValidate=false] Whether to skip validation when
-   * the Record instances are created.
+   * @param {boolean} [opts.noValidate={@link Mapper#noValidate}] See {@link Mapper#noValidate}.
    * @returns {Record|Record[]} The Record instance or Record instances.
    * @since 3.0.0
    */
@@ -1308,7 +1347,7 @@ export default Component.extend({
     if (!utils.isObject(props)) {
       throw utils.err(`${DOMAIN}#createRecord`, 'props')(400, 'array or object', props)
     }
-    const recordClass = this.recordClass
+    const RecordCtor = this.recordClass
     const relationList = this.relationList || []
     relationList.forEach(function (def) {
       const relatedMapper = def.getRelation()
@@ -1321,7 +1360,10 @@ export default Component.extend({
       }
     })
     // Check to make sure "props" is not already an instance of this Mapper.
-    return recordClass ? (props instanceof recordClass ? props : new recordClass(props, opts)) : props // eslint-disable-line
+    if (RecordCtor && (!(props instanceof RecordCtor))) {
+      return new RecordCtor(props, opts)
+    }
+    return props
   },
 
   /**
@@ -2088,6 +2130,7 @@ export default Component.extend({
    * @param {boolean} [opts.adapter={@link Mapper#defaultAdapter}] Name of the
    * adapter to use.
    * @param {boolean} [opts.notify={@link Mapper#notify}] See {@link Mapper#notify}.
+   * @param {boolean} [opts.noValidate={@link Mapper#noValidate}] See {@link Mapper#noValidate}.
    * @param {boolean} [opts.raw={@link Mapper#raw}] See {@link Mapper#raw}.
    * transaction.
    * @returns {Promise} Resolves with the updated record. Rejects if the record
@@ -2179,6 +2222,7 @@ export default Component.extend({
    * @param {boolean} [opts.adapter={@link Mapper#defaultAdapter}] Name of the
    * adapter to use.
    * @param {boolean} [opts.notify={@link Mapper#notify}] See {@link Mapper#notify}.
+   * @param {boolean} [opts.noValidate={@link Mapper#noValidate}] See {@link Mapper#noValidate}.
    * @param {boolean} [opts.raw={@link Mapper#raw}] See {@link Mapper#raw}.
    * @returns {Promise} Resolves with the update records, if any.
    * @see query
@@ -2263,6 +2307,7 @@ export default Component.extend({
    * @param {boolean} [opts.adapter={@link Mapper#defaultAdapter}] Name of the
    * adapter to use.
    * @param {boolean} [opts.notify={@link Mapper#notify}] See {@link Mapper#notify}.
+   * @param {boolean} [opts.noValidate={@link Mapper#noValidate}] See {@link Mapper#noValidate}.
    * @param {boolean} [opts.raw={@link Mapper#raw}] See {@link Mapper#raw}.
    * @returns {Promise} Resolves with the updated records. Rejects if any of the
    * records could be found.
@@ -2395,7 +2440,7 @@ export default Component.extend({
  *
  * @example <caption>Mapper.extend</caption>
  * // Normally you would do: import {Mapper} from 'js-data'
- * const JSData = require('js-data@3.0.0-beta.7')
+ * const JSData = require('js-data@3.0.0-beta.10')
  * const {Mapper} = JSData
  * console.log('Using JSData v' + JSData.version.full)
  *